pidfilechecker advanced

+++ checkers/pidfilechecker.py
@@ -3,20 +3,41 @@
 
 import os
 import logging
+import errno
 
 class PidFileChecker(object):
 
     def __init__(self, **kwargs):
         self.pid_file = kwargs['pid_file']
-        
-    def check_pid(self, pid):        
-        """ Check For the existence of a unix pid. """
+            
+    def pid_exists(self, pid):
+        """Check whether pid exists in the current process table.
+        UNIX only.
+        """
+        if pid < 0:
+            return False
+        if pid == 0:
+            # According to "man 2 kill" PID 0 refers to every process
+            # in the process group of the calling process.
+            # On certain systems 0 is a valid PID but we have no way
+            # to know that in a portable fashion.
+            raise ValueError('invalid PID 0')
         try:
             os.kill(pid, 0)
-        except OSError:
-            return False
+        except OSError as err:
+            if err.errno == errno.ESRCH:
+                # ESRCH == No such process
+                return False
+            elif err.errno == errno.EPERM:
+                # EPERM clearly means there's a process to deny access to
+                return True
+            else:
+                # According to "man 2 kill" possible error values are
+                # (EINVAL, EPERM, ESRCH)
+                raise
         else:
             return True
+            
         
     def is_active(self):
         with open(self.pid_file, 'r') as file:
@@ -25,7 +46,7 @@ class PidFileChecker(object):
             try:
                 pid = int(line)
                 logging.debug("read {} from file {}".format(pid, self.pid_file))
-                return self.check_pid(pid)
+                return self.pid_exists(pid)
             except ValueError as ve:
                 logging.error("ValueError error ({0}): {1}".format(e.errno, e.strerror))
                 return False
